[reorganization]:Move delete blockers logic to its own app
Fixes #7362
