+++ specifyweb/frontend/js_src/lib/hooks/__tests__/useDeleteBlockers.test.tsx
@@ -31,7 +31,10 @@ const deleteBlockerResponse = [
   },
 ];
 
-overrideAjax(`/api/delete_blockers/agent/${agentId}/`, deleteBlockerResponse);
+overrideAjax(
+  `/delete_blockers/delete_blockers/agent/${agentId}/`,
+  deleteBlockerResponse
+);
 overrideAjax(
   '/stored_query/ephemeral/',
   {
