+++ specifyweb/specify/tests/test_delete_blockers.py
@@ -4,7 +4,7 @@ import json
 from specifyweb.backend.trees.tests.test_trees import GeographyTree
 
 def _url(obj):
-    return f"/api/delete_blockers/{obj._meta.model_name}/{obj.id}/"
+    return f"/delete_blockers/delete_blockers/{obj._meta.model_name}/{obj.id}/"
 
 class TestDeleteBlockers(GeographyTree):
