+++ specifyweb/specify/views.py
@@ -19,7 +19,7 @@ from django.views.decorators.cache import cache_control
 from django.views.decorators.http import require_POST, require_http_methods
 from specifyweb.specify.api import get_model
 
-from specifyweb.middleware.general import require_GET, require_http_methods
+from specifyweb.middleware.general import require_http_methods
 from . import api, models as spmodels
 from .specify_jar import specify_jar, specify_jar_path
 from .uiformatters import get_uiformatter_by_name
@@ -89,35 +89,6 @@ def raise_error(request):
     """
     raise Exception('This error is a test. You may now return to your regularly '
                     'scheduled hacking.')
-
-
-@login_maybe_required
-@require_http_methods(['GET', 'HEAD'])
-def delete_blockers(request, model, id):
-    """Returns a JSON list of fields on <model> that point to related
-    resources which prevent the resource <id> of that model from being
-    deleted.
-    """
-    obj = api.get_object_or_404(model, id=int(id))
-    using = router.db_for_write(obj.__class__, instance=obj)
-    collector = Collector(using=using)
-    collector.delete_blockers = []
-    collector.collect([obj])
-    result = flatten([
-        [
-            {
-                'table': sub_objs[0].__class__.__name__,
-                'field': field.name,
-                'ids': [sub_obj.id for sub_obj in sub_objs]
-            }
-        ] for field, sub_objs in collector.delete_blockers
-    ])
-    return http.HttpResponse(api.toJson(result), content_type='application/json')
-
-
-def flatten(l):
-    return [item for sublist in l for item in sublist]
-
 
 @login_maybe_required
 @require_http_methods(['GET', 'HEAD'])
